requirements.txt
openai
Flask
faiss-cpu